Final Day Hackathon/ai/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
langchain>=0.1.9
langchain-openai>=0.0.8
langgraph>=0.0.26
openai>=1.12.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
cssutils>=2.9.0
typing-extensions>=4.9.0
chromadb>=0.4.22
tiktoken>=0.5.3 